SpaceShip Game
The game includes a group of randomly generated asteroids, and a space
ship that can be controlled by the mouse. The space ship will fire to
shoot asteroids when mouse is pressed. When the fire hit asteroid(s),
score will be given to player. The space ship will be destroyed when
colliding with asteroids. The total count of life is 3. The total
playing time is limited at 10 seconds.